requirements.txt
celery~=5.0.5
redis~=3.5.3
django-celery-results~=2.0.1
django-better-admin-arrayfield~=1.4.2